Split doctor runner internals
